PAAS-584: elasticsearch backup restore
